% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_Licor_model_version.R
\name{detect_Licor_model_version}
\alias{detect_Licor_model_version}
\title{Detect the Licor model version}
\usage{
detect_Licor_model_version(Licor_fn, n_header_rows_min = 100)
}
\arguments{
\item{Licor_fn}{Licor filename to read}

\item{n_header_rows_min}{Number of lines to read, should be more than the number of header lines}
}
\value{
Licor_model includes \code{model} number or "Not Detected"
}
\description{
Detect the Licor model version
}
