% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Licor_6400.R
\name{read_Licor_6400}
\alias{read_Licor_6400}
\title{read Licor 6400 file}
\usage{
read_Licor_6400(Licor_fn)
}
\arguments{
\item{Licor_fn}{Licor filename to read}
}
\value{
Licor A list including a \code{header} list and \code{data} tibble
}
\description{
Read Licor file
Find the row where the column headers and data begin.
Separate any remarks into its own column.
}
